% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_dict}
\alias{codon_dict}
\title{Key and Values for codon translation to amino acids}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_dict
}
\description{
Translates from single codon to amino acid
}
\keyword{datasets}
