% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_to_codons.R
\name{format_to_codons}
\alias{format_to_codons}
\title{Divide RNA sequence into codons.}
\usage{
format_to_codons(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{RNA sequence as a string}

\item{start}{start position of the RNA sequence}
}
\value{
a vector of codons
}
\description{
Divide RNA sequence into codons.
}
\examples{
rna_seq <- "AUCUUACUG"   # examples rna sequence
format_to_codons(rna_seq, start = 1)

}
