% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{generate_dna}
\usage{
generate_dna(length_of_dna)
}
\arguments{
\item{length_of_dna}{Length of dna string to generate}
}
\value{
Returns a random generated string of DNA of the given length
}
\description{
generate a dna string by the given length
}
\examples{

#Generates a dba string of length 10
generate_dna(10)

#Generates a dba string of length 20
generate_dna(20)

}
