% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_code}
\alias{genetic_code}
\title{Genetic Code Data}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
genetic_code
}
\description{
This is the genetic code for AA from DNA sequence
}
\keyword{datasets}
