% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_funcs.R
\name{plot_aa_dist}
\alias{plot_aa_dist}
\title{Visualize the distribution of amino acids in a protein sequence string}
\usage{
plot_aa_dist(data)
}
\arguments{
\item{data}{A character string with amino acids without spaces}
}
\value{
The function will return a ggplot plot object visualizing the count
distribution of the amino acids from the input string
}
\description{
Visualize the distribution of amino acids in a protein sequence string
}
\examples{
# As an example here is a randomly generated amino acid string
aa_string <- "SQRMHLPYRFHHPHPLCGKSNRWEAVFVIHEWQGSESTFPKQEWFFHQQSKVGTRNKPWKDG"

# Plot the count distribution of the amino acids in the string
plot_aa_dist(aa_string)

}
