% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base_freq.R
\name{plot_base_seq}
\alias{plot_base_seq}
\title{Plot Nucleotide Frequence from Sequence}
\usage{
plot_base_seq(seq)
}
\arguments{
\item{seq:}{sequence of ribonucleotides or desoxirebonucleotides as a string.}
}
\value{
bar plot with the absolute frequencies (counts) of the input sequence.
}
\description{
Plot Nucleotide Frequence from Sequence
}
\examples{
plot_base_seq("AGCTAGCTAGCTCGGTTTCGTAGCTCATGC")
plot_base_seq("AUCGAUGGCUAGGCUGUAGCGAUACGUAGCGAUGCUA")

}
