% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_to_U.R
\name{T_to_U}
\alias{T_to_U}
\title{Threonine to uracil}
\usage{
T_to_U(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{The DNA sequence that should have T converted to U.}
}
\value{
The RNA sequence (the DNA sequence with T converted to U)
}
\description{
Threonine to uracil
}
\examples{
DNA <- "ATCA"
RNA <- T_to_U(DNA)
RNA
}
