% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{Translates a vector of codons to amino acid string}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{character vector containing codons of length 3}
}
\value{
Amino acid string translated from codons
}
\description{
Translates a vector of codons to amino acid string
}
\examples{
test_codons <- c("CAA", "CGA", "CUG", "CCG") # example character vector of codons
translate_codons(test_codons)
}
